+++ instapy/login_util.py
@@ -234,7 +234,8 @@ def login_user(
     assert username, "Username not provided"
     assert password, "Password not provided"
 
-    # Hotfix - this check crashes more often than not -- plus in not necessary, I can verify my own connection
+    # Hotfix - this check crashes more often than not -- plus in not necessary,
+    # I can verify my own connection
     if want_check_browser:
         if not check_browser(browser, logfolder, logger, proxy_address):
             return False