+++ CHANGELOG.md
@@ -9,6 +9,9 @@ _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added 
 - Use random tag list for `session.like_by_tags`
 
+### Changed
+- General log rotation, gecko driver log in user directory, comments in 80 chars
+
 ### Fixed
 - Unfollowing of users that haven't posted anything
 - `get_links` xpath for yet another change